requirements.txt
# Core ML Frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
accelerate>=0.25.0

# Vision-Language Models
qwen-vl-utils>=0.0.3

# OCR Engines
paddlepaddle>=2.5.0
paddleocr>=2.7.0
easyocr>=1.7.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# NLP & Utilities
sentencepiece>=0.1.99
protobuf>=3.20.0
regex>=2023.10.3

# Data Processing
pandas>=2.0.0
tqdm>=4.65.0

# API & Network
requests>=2.31.0
aiohttp>=3.9.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development Tools
ipython>=8.12.0
jupyter>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Model Download
huggingface-hub>=0.19.0

# Flash Attention (optional, for speed)
# flash-attn>=2.3.0  # Requires CUDA, install manually if needed

# Other Dependencies
scikit-learn>=1.3.0
scipy>=1.11.0
